Main Menu added
The main menu has been added with most functinality, incuding play to start the game and quit to exit the game.
A level named mainmenu has been added to start the game in and display the main menu. the players movment will also be disabled when in the main menu.