Menu and Pasue Menu Changes
The Main Menu and the Pause Menu now have a rough design, some parts may change later on.

A options menu has also been added to the Main Menu and the Pause Menu. This has no fucntiality as of yet, but you are able to toggle between the menus now with no problumes.

Tommrows work I will be attempting to add the animal models to the prototype.