Egg Model Added
This Egg model may need cahngeing but let me know if it is looking all good when rigged up to the player. If there is any changes that need to be made with the egg let me know.

Also the egg mesh is a bit small but can be adjusted with the scale tool when applied to the player BP editor.

I will make sure to inlarge the scales in Maya for the next models to avoid this.