Added Fence
The fence mesh has been added. Also to note there was another issue when importing the mesh over. the issue was that the mesh was plit into two parts, but as a solution i had just made a BP for the fence and reconnected them, hope that will help when buiding.